requirements.txt
python-telegram-bot>=20.0
transformers>=4.30.0
torch>=2.0.0
flask